Split code generators and rename
